requirements/base.txt
django==4.2.5
djangorestframework==3.14.0 
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
django-environ==0.11.2
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.9.1
psycopg2-binary==2.9.5
pytz==2023.3
tzdata==2023.3